+++ conftest.py
@@ -36,7 +36,6 @@ def global_test_env(request):
     env = create_env(tmp_dir, node_file, account_file, init_chain, install_dependency, install_supervisor)
 
     env.deploy_all()
-
     yield env
 
     # todo
@@ -48,12 +47,11 @@ def pytest_runtest_makereport(item, call):
     # execute all other hooks to obtain the report object
     outcome = yield
     rep = outcome.get_result()
-    log.info("try download log")
     # we only look at actual failing test calls, not setup/teardown
     if rep.when == "call" and not rep.passed:
         # download log in here
         try:
-            log.info(rep.global_test_env.chain_id)
-            # item.global_test_env.backup_all_logs(item.name)
+            if 'global_test_env' in item.fixturenames:
+                log.info(item.funcargs["global_test_env"].backup_all_logs(item.name))
         except Exception as e:
             log.info("这个失败没有生成日志:{}".format(e))
