+++ pytest.ini
@@ -6,3 +6,6 @@ markers =
     P3: P3
     P4: P4
     SYNC: SYNC
+log_cli = true
+log_cli_format = %(asctime)s %(levelname)s %(message)s
+log_cli_date_format = %Y-%m-%d %H:%M:%S
