--- tests/lib/keyword.py
@@ -1,28 +1,0 @@
-
-UnStakeFreezeRatio = "UnStakeFreezeRatio"
-
-VersionProposalActiveConsensusRounds = "VersionProposalActive_ConsensusRounds"
-
-VersionProposalSupportRate = "VersionProposal_SupportRate"
-
-TextProposalVoteDurationSeconds = "TextProposalVote_DurationSeconds"
-
-TextProposalVoteRate = "TextProposal_VoteRate"
-
-TextProposalSupportRate = "TextProposal_SupportRate"
-
-CancelProposalVoteRate = "CancelProposal_VoteRate"
-
-CancelProposalSupportRate = "CancelProposal_SupportRate"
-
-StakeThreshold = "StakeThreshold"
-
-ValidatorCount = "ValidatorCount"
-
-EpochValidatorNum = "EpochValidatorNum"
-
-Period = "period"
-
-Amount = "amount"
-
-ChainId = "chainId"
